+++ docs/src/dataDisplay/personCoreMilestones/personCoreMilestones.jsx
@@ -10,6 +10,7 @@ import ComponentVersionIdentifier from '../../global/componentVersionIdentifier'
 import Example from '../../global/example';
 import ExampleAdultFemale from './examples/exampleAdultFemale';
 import ExampleAdultMale from './examples/exampleAdultMale';
+import ExampleInteractive from './examples/exampleInteractive';
 import ExampleChild from './examples/exampleChild';
 import ExampleDisable from './examples/exampleDisable';
 import ExampleStudent from './examples/exampleStudent';
@@ -128,6 +129,25 @@ function DocsPersonCoreMilestones(props) {
                         rawCode={require('!!raw-loader!./examples/exampleDisable').default}
                     >
                         <ExampleDisable />
+                    </Example>
+
+                    <Heading
+                        anchorLink="disable-popover"
+                        variant="h2"
+                    >
+                        Interactive Person Core Milestones
+                    </Heading>
+
+                    <Typography
+                        variant="body1"
+                    >
+                        Go ahead, give it a whirl
+                    </Typography>
+
+                    <Example
+                        rawCode={require('!!raw-loader!./examples/exampleInteractive').default}
+                    >
+                        <ExampleInteractive />
                     </Example>
                 </MarkdownContainer>